Clean up and add new sources
