+++ manifest.yml
@@ -16,4 +16,5 @@ sources:
         neighborhood_councils: http://geohub.lacity.org/datasets/674f80b8edee4bf48551512896a1821d_0
         council_districts: http://geohub.lacity.org/datasets/76104f230e384f38871eb3c4782f903d_13
         community_plan_areas: http://geohub.lacity.org/datasets/f4e830d3882d457bb9d9241cdff0b10a_0
-        zoning: http://geohub.lacity.org/datasets/49ad06a6b8c945debbbea865b1832ee2_0       
+        zoning: http://geohub.lacity.org/datasets/49ad06a6b8c945debbbea865b1832ee2_0
+        metrolink_stations: http://geohub.lacity.org/datasets/6f6c4677365b4418bd585db2ef8e201f_186
